requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
